+++ src/ggg.rs
@@ -464,11 +464,11 @@ pub struct FeatureVariation<'a> {
 impl<'a> FeatureVariation<'a> {
     /// Evaluates variation using specified `coordinates`.
     ///
-    /// Note: coordinates should be converted from fixed point 2.14 to i32
+    /// Note: coordinates should be converted from fixed point 2.14 to i16
     /// by multiplying each coordinate by 16384.
     ///
     /// Number of `coordinates` should be the same as number of variation axes in the font.
-    pub fn evaluate(&self, coordinates: &[i32]) -> bool {
+    pub fn evaluate(&self, coordinates: &[i16]) -> bool {
         for condition in try_opt_or!(self.condition_set(), false) {
             if !condition.evaluate(coordinates) {
                 return false;
@@ -534,9 +534,9 @@ impl<'a> Iterator for ConditionSet<'a> {
 
 
 impl Condition {
-    fn evaluate(&self, coordinates: &[i32]) -> bool {
+    fn evaluate(&self, coordinates: &[i16]) -> bool {
         let coord = coordinates.get(self.axis_index() as usize).cloned().unwrap_or(0);
-        self.filter_range_min_value() as i32 <= coord && coord <= self.filter_range_max_value() as i32
+        self.filter_range_min_value() <= coord && coord <= self.filter_range_max_value()
     }
 }
